+++ jkube-kit/enricher/generic/src/test/java/org/eclipse/jkube/enricher/generic/DefaultNamespaceEnricherTest.java
@@ -13,11 +13,13 @@
  */
 package org.eclipse.jkube.enricher.generic;
 
-import io.fabric8.kubernetes.api.model.ConfigMapBuilder;
 import io.fabric8.kubernetes.api.model.KubernetesListBuilder;
 import io.fabric8.kubernetes.api.model.Namespace;
 import io.fabric8.kubernetes.api.model.NamespaceBuilder;
+import io.fabric8.kubernetes.api.model.apps.Deployment;
+import io.fabric8.kubernetes.api.model.apps.DeploymentBuilder;
 import io.fabric8.openshift.api.model.Project;
+import io.fabric8.openshift.api.model.ProjectBuilder;
 import mockit.Expectations;
 import mockit.Mocked;
 import org.eclipse.jkube.kit.config.resource.PlatformMode;
@@ -30,6 +32,7 @@ import java.util.Collections;
 import java.util.Properties;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.Assert.assertNull;
 
 public class DefaultNamespaceEnricherTest {
 
@@ -108,18 +111,71 @@ public class DefaultNamespaceEnricherTest {
   }
 
   @Test
-  public void enrichWithPropertiesInKubernetesShouldAddNamespace() {
+  public void enrichWithPropertiesInKubernetesShouldAddNamespaceWithStatus() {
     // Given
     properties.put("jkube.enricher.jkube-namespace.namespace", "example");
     final KubernetesListBuilder klb = new KubernetesListBuilder();
-    klb.addToItems(new NamespaceBuilder()
-        .editOrNewMetadata().withName("name").withNamespace("to-be-overwritten").endMetadata()
-        .editOrNewStatus().withPhase("active").endStatus().build());
+    Namespace namespace = new NamespaceBuilder()
+            .editOrNewMetadata().withName("name").withNamespace("to-be-overwritten").endMetadata()
+            .editOrNewStatus().withPhase("active").endStatus().build();
+    Deployment deployment = new DeploymentBuilder().withNewMetadata().withName("d1").endMetadata().build();
+    klb.addToItems(namespace, deployment);
     // When
     new DefaultNamespaceEnricher(context).enrich(PlatformMode.kubernetes, klb);
     // Then
+    assertThat(klb.build().getItems()).hasSize(2);
+    assertThat(klb.build().getItems().get(1))
+        .hasFieldOrPropertyWithValue("metadata.namespace", "example");
+  }
+
+  @Test
+  public void enrichWithPropertiesInKubernetesShouldAddProjectWithStatus() {
+    // Given
+    final KubernetesListBuilder klb = new KubernetesListBuilder();
+    klb.addToItems(new ProjectBuilder()
+            .withNewMetadata().withName("name").endMetadata()
+            .withNewStatus().withPhase("active").endStatus().build());
+    // When
+    new DefaultNamespaceEnricher(context).enrich(PlatformMode.openshift, klb);
+    // Then
     assertThat(klb.build().getItems()).hasSize(1);
     assertThat(klb.build().getItems().iterator().next())
-        .hasFieldOrPropertyWithValue("metadata.namespace", "example");
+            .hasFieldOrPropertyWithValue("metadata.namespace", null);
+  }
+
+  @Test
+  public void enrichWithNamespaceFragmentWithNoStatus() {
+    // Given
+    final KubernetesListBuilder kubernetesListBuilder = new KubernetesListBuilder();
+    kubernetesListBuilder.addToItems(new NamespaceBuilder()
+            .withNewMetadata().withName("test-jkube").endMetadata()
+            .build());
+
+    // When
+    new DefaultNamespaceEnricher(context).enrich(PlatformMode.kubernetes, kubernetesListBuilder);
+
+    // Then
+    assertThat(kubernetesListBuilder.build().getItems()).hasSize(1);
+    assertThat(kubernetesListBuilder.build().getItems().iterator().next())
+            .hasFieldOrPropertyWithValue("metadata.name", "test-jkube");
+    assertNull(kubernetesListBuilder.build().getItems().get(0).getMetadata().getNamespace());
+  }
+
+  @Test
+  public void enrichWithOpenShiftProjectFragmentWithNoStatus() {
+    // Given
+    final KubernetesListBuilder kubernetesListBuilder = new KubernetesListBuilder();
+    kubernetesListBuilder.addToItems(new ProjectBuilder()
+            .withNewMetadata().withName("test-jkube").endMetadata()
+            .build());
+
+    // When
+    new DefaultNamespaceEnricher(context).enrich(PlatformMode.openshift, kubernetesListBuilder);
+
+    // Then
+    assertThat(kubernetesListBuilder.build().getItems()).hasSize(1);
+    assertThat(kubernetesListBuilder.build().getItems().iterator().next())
+            .hasFieldOrPropertyWithValue("metadata.name", "test-jkube");
+    assertNull(kubernetesListBuilder.build().getItems().get(0).getMetadata().getNamespace());
   }
 }